// prototype/reading-queue-reference/frontend/src/generated/public-api/model/problem.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Reading Queue Testbed Public API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';

export const problemStatusMin = 0;


export const Problem = zod.object({
  "detail": zod.string(),
  "status": zod.int().min(problemStatusMin),
  "title": zod.string(),
  "traceId": zod.string(),
  "type": zod.string()
})

export type Problem = zod.input<typeof Problem>;
export type ProblemOutput = zod.output<typeof Problem>;